move settings to pages
